refactor: move SettingsModal to components/Modals folder
Moves SettingsModal from components/BuildQuote/Modals/ to
components/Modals/ for better organization as a shared modal component.
