+++ app/components/UI/Ramp/routes.tsx
@@ -4,7 +4,7 @@ import Routes from '../../../constants/navigation/Routes';
 import TokenSelection from './components/TokenSelection';
 import BuildQuote from './components/BuildQuote';
 import UnsupportedTokenModal from './components/UnsupportedTokenModal';
-import SettingsModal from './components/BuildQuote/Modals/SettingsModal';
+import SettingsModal from './components/Modals/SettingsModal';
 
 const RootStack = createStackNavigator();
 const Stack = createStackNavigator();
